+++ test/fakeGlobal.ts
@@ -1,8 +1,0 @@
-(global as any).tinymce = {
-  init: (x: any) => {
-    return;
-  }
-};
-(global as any).requestAnimationFrame = (callback: any) => {
-  setTimeout(callback, 0);
-};